deleted latest.js from login page
