+++ src/pages/login/index.js
@@ -1,11 +1,9 @@
 import LoginPage from "./Login";
-import Latest from "./Latest";
 
 const Login = () => {
  return (
   <div>
    <LoginPage />
-   <Latest />
   </div>
  );
 };
